#37 Refactored the headerStyle property of the navigators.
