+++ chicken/src/config/routes/authenticated/AuthenticatedScanNavigator.js
@@ -1,6 +1,6 @@
 import { createStackNavigator } from 'react-navigation';
 
-import { colors, globalStyles } from '../../globalStyles';
+import { globalStyles } from '../../globalStyles';
 
 import ScanScreen from '../../../screens/authenticated/scanTab/ScanScreen';
 
@@ -10,9 +10,7 @@ const AuthenticatedScanNavigator = createStackNavigator({
     navigationOptions: () => ({
       title: 'Scan',
       headerTitleStyle: globalStyles.navigationHeaderTitle,
-      headerStyle: {
-        backgroundColor: colors.navigationHeaderBackground
-      }
+      headerStyle: globalStyles.navigationHeaderStyle
     })
   }
 });
